+++ app/src/main/java/com/xiaomiquan/entity/bean/kline/KLineBean.java
@@ -1,16 +1,22 @@
 package com.xiaomiquan.entity.bean.kline;
 
+import java.math.BigDecimal;
+
 /**
  * Created by loro on 2017/2/8.
  */
 public class KLineBean {
     public String date;
     public long timestamp;
-    public float open;
-    public float close;
-    public float high;
-    public float low;
-    public float volume;
+
+    public BigDecimal open;
+    public BigDecimal close;
+    public BigDecimal high;
+    public BigDecimal low;
+    public BigDecimal volume;
+
+
+
 
     @Override
     public String toString() {
